Rename Prepper into makeSuite
